client: Respect O_SYNC, O_DSYNC
If the file is opened with O_SYNC or O_DSYNC, we need to
flush cached data (and metadata for O_SYNC) on a write.
This patch adds a flags field directly to the file handle
struct, and replaces the append boolean with the use of the
flags field directly.

Signed-off-by: Sam Lang <sam.lang@inktank.com>